requirements.txt
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==4.0.2
discord-py-interactions==4.3.1
discord-py-slash-command==1.1.0
discord.py